requirements.txt
# Core ML/AI Libraries
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0

# Data Processing and Augmentation
numpy>=1.21.0
opencv-python>=4.7.0
albumentations>=1.3.0
Pillow>=9.0.0

# Machine Learning Utilities
scikit-learn>=1.2.0

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0

# Web Interface
gradio>=4.0.0

# Configuration and Utilities
PyYAML>=6.0
tqdm>=4.64.0
pathlib

# Development and Testing (optional)
pytest>=7.0.0
jupyter>=1.0.0

# GPU Support (uncomment if using CUDA)
# torch-audio>=2.0.0
# torchtext>=0.15.0

# Additional utilities
pandas>=1.5.0
plotly>=5.0.0
